refactor(expressive): rebuild BottomTabBar on HorizontalFloatingToolbar
Move the BottomTabBar component from `ui` to `expressive` so it can sit
on top of Material3's HorizontalFloatingToolbar (matches how the Teya
Business App's TeyaBottomBar is structured today).

What changed:
- HorizontalFloatingToolbar wraps the items, with a 4dp content padding
  that mirrors the Figma "Tabs Slot" inner padding spec.
- LemonadeShadow.Xlarge is reapplied via lemonadeShadow() so the
  Figma shadow-xlarge drop is preserved on top of the toolbar surface.
- Item interactions go through clickable + the project's
  LocalEffects.current.interactionIndication, restoring the press/hover
  ripple that selectable + the default indication wasn't producing.
- Drop the explicit max-width and selectable wiring; the component now
  follows the same interaction pattern as Tile/Button/Chip.
- BottomTabBarItem gains an optional selectedIcon to swap in a paired
  variant (e.g. Wallet -> WalletSolid) while selected.
- BottomTabBarDisplay reuses its existing sections to demo the
  selectedIcon swap across all sample sizes.

Addresses Copilot review feedback on indication / interactionSource
handling on the original ui-module implementation.
